Apply copier template
