+++ .github/workflows/main.yml
@@ -1,3 +1,6 @@
+# Do not edit this file directly
+# Autogenerated by gh:AmpelAstro/copier-ampel
+
 name: ci
 
 on:
@@ -16,16 +19,32 @@ on:
 
 jobs:
   ci:
-    uses: AmpelProject/Ampel-interface/.github/workflows/ci.yml@e880b6e4cc9827cb4a85a06d00cb36738d9dda93 # ci-py12-v3
-    secrets:
-      PYPI_API_TOKEN: ${{ secrets.PYPI_API_TOKEN }}
+    uses: AmpelProject/Ampel-interface/.github/workflows/ci.yml@c855e93ae95c955ba292f88413455462398a8c9a # ci-py12-v10
     with:
-      mongo: false
-      lint: true
+      mongo: true
       format: true
+      lint: true
       # renovate: datasource=python-version depName=python
       python-version: "3.14.1"
+      # renovate: datasource=pypi depName=ruff versioning=pep440
+      ruff-version: "0.14.7"
       # renovate: datasource=pypi depName=poetry versioning=pep440
       poetry-version: "2.2.1"
-      # renovate: datasource=pypi depName=ruff
-      ruff-version: "0.14.7"
+  # Need to keep this step separate until pypi trusted publishing supports reusable workflows: https://github.com/pypi/warehouse/issues/11096
+  pypi:
+    needs: ci
+    runs-on: ubuntu-24.04
+    # NB: outputs are always strings; explicitly parse to get a boolean
+    if: ${{ fromJSON(needs.ci.outputs.should_publish) }}    
+    permissions:
+      # IMPORTANT: this permission is mandatory for trusted publishing, but
+      # should NOT be granted anywhere else!
+      id-token: write
+
+    steps:
+      - name: Download built artifact to dist/
+        uses: actions/download-artifact@v4
+        with:
+          name: Packages
+          path: dist
+      - uses: pypa/gh-action-pypi-publish@release/v1
