+++ .gitignore
@@ -1,12 +1,7 @@
+.venv/
 *.egg-info
-*.swp
-*.swo
-*.swn
 __pycache__
 .pytest_cache
-.DS_Store
 .mypy_cache
-dist/
-build
-public
-.vscode/
+.vscode
+dist
