+++ .pre-commit-config.yaml
@@ -4,7 +4,7 @@
 repos:
 - repo: https://github.com/astral-sh/ruff-pre-commit
   # renovate: datasource=pypi depName=ruff versioning=pep440
-  rev: "v0.14.7"
+  rev: v0.14.8
   hooks:
     - id: ruff
       args: [--fix]
@@ -14,4 +14,10 @@ repos:
   rev: "2.2.1"
   hooks: 
     - id: poetry-check
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v4.3.0
+  hooks:
+      # Prevent committing inline conflict markers
+      - id: check-merge-conflict
+        args: [--assume-in-merge]
 fail_fast: false
