+++ CONTRIBUTING.md
@@ -1,0 +1,11 @@
+
+# Style guidelines
+
+## Use pre-commit
+
+Pre-commit hooks check your commits for obvious mistakes like invalid
+syntax in Python or JSON files.  [Install pre-commit on your
+machine](https://pre-commit.com/#installation), then activate the
+hooks with `pre-commit install`.  The configured hooks will then
+run before every commit, either fixing the files to be committed
+or telling you what's wrong.
