+++ ampel/conf/ampel-lsst/ampel.yml
