+++ tests/test_ingestion.py
@@ -1,9 +1,12 @@
+from itertools import cycle
 from pathlib import Path
 
 import fastavro
 import pytest
 import yaml
 
+from ampel.abstract.AbsAlertFilter import AbsAlertFilter, AmpelAlertProtocol
+from ampel.abstract.AbsAlertLoader import AbsAlertLoader
 from ampel.alert.AlertConsumer import AlertConsumer
 from ampel.base.AuxUnitRegister import AuxUnitRegister
 from ampel.dev.DevAmpelContext import DevAmpelContext
@@ -11,16 +14,38 @@ from ampel.lsst.alert.LSSTAlertSupplier import LSSTAlertSupplier
 from ampel.model.UnitModel import UnitModel
 
 
+class MockAlertLoader(AbsAlertLoader):
+    alerts: list[dict]
+
+    def __init__(self, **kwargs) -> None:
+        super().__init__(**kwargs)
+        self._it = iter(self.alerts)
+
+    def _add_metadata(self, alert: dict) -> dict:
+        alert["__kafka"] = {"alertId": alert["alertId"]}
+        return alert
+
+    def __next__(self):
+        return self._add_metadata(next(self._it))
+
+
+class MockFilter(AbsAlertFilter):
+    pattern: list[bool]
+
+    def __init__(self, **kwargs) -> None:
+        super().__init__(**kwargs)
+        self._cycle = cycle(self.pattern)
+
+    def process(self, alert: AmpelAlertProtocol) -> None | bool | int:  # noqa: ARG002
+        return next(self._cycle)
+
+
 @pytest.fixture
 def alert_consumer(mock_context: DevAmpelContext) -> AlertConsumer:
-    with (
-        Path(__file__).parent / "test-data" / "elasticc-consumer.yml"
-    ).open() as f:
+    with (Path(__file__).parent / "test-data" / "elasticc-consumer.yml").open() as f:
         model = UnitModel(**yaml.safe_load(f))
     # alerts from a single diaObject
-    with (Path(__file__).parent / "test-data" / "11290844.avro").open(
-        "rb"
-    ) as f:
+    with (Path(__file__).parent / "test-data" / "11290844.avro").open("rb") as f:
         alerts = list(fastavro.reader(f))[:2]
     model.config["supplier"]["config"]["alert_identifier"] = "alertId"
     model.config["supplier"]["config"]["loader"] = UnitModel(
@@ -33,6 +58,11 @@ def alert_consumer(mock_context: DevAmpelContext) -> AlertConsumer:
     model.config["directives"][1]["filter"] = UnitModel(
         unit="MockFilter", config={"pattern": [False, True]}
     ).dict()
+
+    for c in "ElasticcLong", "ElasticcShort":
+        mock_context.add_channel(c)
+    mock_context.register_unit(MockAlertLoader)
+    mock_context.register_unit(MockFilter)
 
     return mock_context.loader.new_context_unit(
         model=model,
@@ -55,15 +85,13 @@ def test_muxer(mock_context: DevAmpelContext, alert_consumer: AlertConsumer):
     # inserts datapoints unique to second alert
     assert alert_consumer.run() == 1
 
-    assert (
-        len(mock_context.db.get_collection("stock").find_one()["channel"]) == 2
-    ), "stock in both channels"
+    assert len(mock_context.db.get_collection("stock").find_one()["channel"]) == 2, (
+        "stock in both channels"
+    )
     assert (
         len(
             docs := list(
-                mock_context.db.get_collection("t2").find(
-                    {"unit": "T2GetDiaObject"}
-                )
+                mock_context.db.get_collection("t2").find({"unit": "T2GetDiaObject"})
             )
         )
         == 1
@@ -97,10 +125,10 @@ def test_duplicate_datapoints(mock_context: DevAmpelContext):
     result in states with unique datapoints.
     """
 
-    with (
-        Path(__file__).parent / "test-data" / "elasticc-consumer.yml"
-    ).open() as f:
+    with (Path(__file__).parent / "test-data" / "elasticc-consumer.yml").open() as f:
         model = UnitModel(**yaml.safe_load(f))
+    for directive in model.config["directives"]:
+        mock_context.add_channel(directive["channel"])
     # alert with duplicated datapoints
     model.config["supplier"]["config"]["loader"] = UnitModel(
         unit="ElasticcDirAlertLoader",
@@ -131,6 +159,4 @@ def test_duplicate_datapoints(mock_context: DevAmpelContext):
     assert processor.run() == 1
 
     assert (t1 := mock_context.db.get_collection("t1").find_one())
-    assert len(t1["dps"]) == len(set(t1["dps"])), (
-        "datapoints in state are unique"
-    )
+    assert len(t1["dps"]) == len(set(t1["dps"])), "datapoints in state are unique"
